# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PolariixApi",
        "description": "Documentation complète de l'API - Tous les endpoints sont accessibles ici",
        "contact": {
            "name": "Support API",
            "email": "contact@votreapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/v1/piliers": {
            "get": {
                "tags": [
                    "Piliers"
                ],
                "summary": "Liste des piliers",
                "description": "Récupère la liste paginée des piliers",
                "operationId": "865b974ff9c0c43242bbcee0073af4b2",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Piliers"
                ],
                "summary": "Créer un pilier",
                "description": "Crée un nouveau pilier",
                "operationId": "9352bcaed5c7c08307668f28e855be6b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Pilier 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description du pilier"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pilier créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/piliers/{pilier}": {
            "get": {
                "tags": [
                    "Piliers"
                ],
                "summary": "Détails d'un pilier par ID",
                "description": "Récupère les détails d'un pilier via son ID.",
                "operationId": "950e0946566e1c500b0ed28e450cadb2",
                "parameters": [
                    {
                        "name": "pilier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pilier trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Piliers"
                ],
                "summary": "Mettre à jour un pilier par ID",
                "description": "Met à jour un pilier existant via son ID.",
                "operationId": "4d24379ee1f8b6997db8340e1d9e447c",
                "parameters": [
                    {
                        "name": "pilier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pilier mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Piliers"
                ],
                "summary": "Supprimer un pilier par ID",
                "description": "Supprime un pilier via son ID.",
                "operationId": "cbe290a73b482895451e7d970686b8cd",
                "parameters": [
                    {
                        "name": "pilier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pilier supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/piliers/slug/{slug}": {
            "get": {
                "tags": [
                    "Piliers"
                ],
                "summary": "Consulter un pilier via son slug (URL publique SEO‑friendly)",
                "description": "Récupère un pilier par son slug URL‑friendly. À utiliser côté front pour des URLs lisibles et pour éviter d'exposer des IDs incrémentaux (anti‑énumération).",
                "operationId": "ed250f135e3bdb3a5876ea1e23b916cd",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pilier trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Piliers"
                ],
                "summary": "Mettre à jour un pilier via son slug (référence URL‑friendly)",
                "description": "Met à jour un pilier en l'identifiant par son slug public, pratique quand seul l'URL publique est connue côté client.",
                "operationId": "3c673d31b8a097d9ac2c2e6c9aca67b9",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pilier mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Piliers"
                ],
                "summary": "Supprimer un pilier via son slug (URL publique)",
                "description": "Supprime un pilier en le ciblant via son slug public, sans exposer l'ID interne.",
                "operationId": "dc7417b29e0afb12a6eefce231e7d205",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pilier supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "title": "Réponse de succès standard",
                "description": "Structure standard pour les réponses réussies",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via JWT Token. Entrez votre token sans le préfixe 'Bearer'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}